go: fix lint on windows and macOS build tags
