+++ go/client/cmd_ctl.go
@@ -4,8 +4,6 @@
 package client
 
 import (
-	"strings"
-
 	"github.com/keybase/cli"
 	"github.com/keybase/client/go/install"
 	"github.com/keybase/client/go/libcmdline"
@@ -46,25 +44,6 @@ func defaultCtlComponents(enable bool) map[string]bool { //nolint
 	components := map[string]bool{}
 	for _, c := range availableCtlComponents {
 		components[c] = enable
-	}
-	return components
-}
-
-// ctlParseArgv returns map with include/exclude components
-func ctlParseArgv(ctx *cli.Context) map[string]bool { //nolint
-	components := defaultCtlComponents(true)
-	if ctx.String("exclude") != "" {
-		excluded := strings.Split(ctx.String("exclude"), ",")
-		for _, exclude := range excluded {
-			components[exclude] = false
-		}
-	}
-	if ctx.String("include") != "" {
-		included := strings.Split(ctx.String("include"), ",")
-		components = defaultCtlComponents(false)
-		for _, include := range included {
-			components[include] = true
-		}
 	}
 	return components
 }
