+++ go/logger/output_windows.go
@@ -73,7 +73,6 @@ const (
 	fgYellow    WORD = 0x0006
 	fgWhite     WORD = 0x0007
 	fgIntensity WORD = 0x0008
-	fgMask      WORD = 0x000F
 
 	bgBlack     WORD = 0x0000
 	bgBlue      WORD = 0x0010
@@ -84,7 +83,6 @@ const (
 	bgYellow    WORD = 0x0060
 	bgWhite     WORD = 0x0070
 	bgIntensity WORD = 0x0080
-	bgMask      WORD = 0x00F0
 )
 
 var codesWin = map[byte]WORD{
