+++ go/tools/winresource/icon.go
@@ -42,7 +42,7 @@ THE SOFTWARE.
 const (
 	rtIcon      = coff.RT_ICON
 	rtGroupIcon = coff.RT_GROUP_ICON
-	rtManifest  = coff.RT_MANIFEST
+	_           = coff.RT_MANIFEST // rtManifest
 )
 
 // on storing icons, see: http://blogs.msdn.com/b/oldnewthing/archive/2012/07/20/10331787.aspx
